move persistDefaultList call to presenter
