+++ app/src/main/java/com/jeffcunningham/twitterlistviewer_android/lists/ListsPresenter.java
@@ -4,4 +4,10 @@ package com.jeffcunningham.twitterlistviewer_android.lists;
  * Created by jeffcunningham on 1/22/17.
  */
 public interface ListsPresenter {
+
+    public void getListMembershipByTwitterUser();
+
+    public void persistDefaultListId(String alias, String listId, String slug, String listName);
+
+
 }
